Extract AreaCard and BerthCard
